SUIT-10728 Resolve CodeClimate JS API issues. A lot of changes related to typescript tests. Splitting big functions in to smalls. Removed duplicated code where it make sense.
