+++ lib/chains/cookieChain.js
@@ -127,9 +127,7 @@ const getComposers = data => {
 
 	if (!data.comparator) {
 		output.push(
-			startWithComposer, endWithComposer, containComposer, equalComposer,
-			matchJSComposer,
-			existComposer
+			startWithComposer, endWithComposer, containComposer, equalComposer, matchJSComposer, existComposer
 		);
 	}
