+++ lib/utils/Queue.js
@@ -1,3 +1,5 @@
+const {curry} = require('ramda');
+
 module.exports = class Queue {
 	constructor(concurrency) {
 		this.functions = [];
@@ -14,17 +16,13 @@ module.exports = class Queue {
 		const threadsCount = this.concurrency > 0 ? this.concurrency : this.functions.length;
 
 		return new Promise(res => {
-			const onPromiseReject = (rej) => {
+			const baseHandler = curry((handler, data) => {
 				this.inProgress--;
-				this.result.push({error: rej});
+				handler(data);
 				next();
-			};
-
-			const onPromiseResolve = (res) => {
-				this.inProgress--;
-				this.result.push({result: res});
-				next();
-			};
+			});
+			const onPromiseReject = baseHandler(error => this.result.push({error}));
+			const onPromiseResolve = baseHandler(result => this.result.push({result}));
 
 			const next = () => {
 				if (!this.functions.length && this.inProgress === 0) {
