+++ lib/utils/socketChainHelper.js
@@ -120,9 +120,8 @@ const applyCountAndDelay = (socketMessage, data) => {
 /**
  * @description Apply default clicks array
  * @param {Object} socketMessage
- * @param {Object} data
  */
-const applyClicks = (socketMessage, data) => {
+const applyClicks = (socketMessage) => {
 	socketMessage.clicks = [{
 		type: 'single',
 		button: 'left',
