+++ lib/utils/socketErrorMessages.js
@@ -95,6 +95,71 @@ function normalizeErrorType(response) {
 	return response.errorType;
 }
 
+function valTemplate(str, isActual) {
+	return colors[isActual ? 'red' : 'green'](
+		`${isActual ? '×' : '~'} ${str} (${isActual ? 'actual' : 'expected'})`
+	);
+}
+
+function getExpressionActualExpected(res, data) {
+	const values = [];
+	const isUntil = 'until' in data;
+	const props = isUntil
+		? path(['until', 'expression'], data)
+		: path(['comparator', 'props'], data);
+
+	// go through expression array
+	for (let i = 0; i < res.expression.length; i++) {
+		const expr = res.expression[i];
+		const prop = props[i];
+		const propName = isUntil ? prop.property : ELEMENT_PROP[prop.name];
+
+		expr.expectedValue !== undefined && values.push(valTemplate(`${propName}: ${expr.expectedValue}`, false));
+		expr.actualValue !== undefined && values.push(valTemplate(`${propName}: ${expr.actualValue}`, true));
+		expr.expectedValue !== undefined && expr.actualValue && values.push('');
+	}
+
+	return values;
+}
+
+function getNetworkActualExpected(res, data) {
+	const values = [];
+	const props = {
+		request: path(['request', 'props'], data) || [],
+		response: path(['response', 'props'], data) || [],
+	};
+
+	// go through errors array
+	for (let i = 0; i < res.errors.length; i++) {
+		const error = res.errors[i];
+		let expected, actual, propName;
+
+		if (error.type === 'noUriFound') {
+			expected = path(['comparator', 'val'], data);
+			actual = t['actualExpected.requestNotMade']();
+			propName = 'url';
+			values.push(valTemplate(`${propName}: ${expected}`, false));
+			values.push(valTemplate(`${propName}: ${actual}`, true));
+			continue;
+		}
+
+		const nameComparator = error.type === 'header' ? error.name : NETWORK_PROP['@' + error.type];
+		const propsField = props[error.message].find(field => field.name === nameComparator);
+
+		if (propsField) {
+			expected = error.type === 'method' ? NETWORK_METHOD[propsField.val] : propsField.val;
+			actual = 'actualValue' in error
+				? error.actualValue
+				: error.reason === 'notFound' ? t['actualExpected.notFound']() : t['actualExpected.unknown']();
+			propName = `${error.message} ${error.type}` + (error.name ? ` "${error.name}"` : '');
+			values.push(valTemplate(`${propName}: ${expected}`, false));
+			values.push(valTemplate(`${propName}: ${actual}`, true));
+		}
+	}
+
+	return values;
+}
+
 /**
  * @description return actualValue and expectedValue depended on server response
  * if response contain top level actualValue and expectedValue - they will be returned.
@@ -107,11 +172,6 @@ function normalizeErrorType(response) {
  */
 function getActualExpectedValues(res, data) {
 	const values = [];
-	const valTemplate = (str, isActual) => {
-		return colors[isActual ? 'red' : 'green'](
-			`${isActual ? '×' : '~'} ${str} (${isActual ? 'actual' : 'expected'})`
-		);
-	};
 
 	// don't display expected value for these error types:
 	if (excludeExpectedValues.includes(res.errorType)) {
@@ -125,56 +185,11 @@ function getActualExpectedValues(res, data) {
 	}
 	// check expression array
 	else if (Array.isArray(res.expression)) {
-		const isUntil = 'until' in data;
-		const props = isUntil
-			? path(['until', 'expression'], data)
-			: path(['comparator', 'props'], data);
-
-		// go through expression array
-		for (let i = 0; i < res.expression.length; i++) {
-			const expr = res.expression[i];
-			const prop = props[i];
-			const propName = isUntil ? prop.property : ELEMENT_PROP[prop.name];
-
-			expr.expectedValue !== undefined && values.push(valTemplate(`${propName}: ${expr.expectedValue}`, false));
-			expr.actualValue !== undefined && values.push(valTemplate(`${propName}: ${expr.actualValue}`, true));
-			expr.expectedValue !== undefined && expr.actualValue && values.push('');
-		}
+		values.push(...getExpressionActualExpected(res, data));
 	}
 	// check networkRequest errors array
 	else if (data.type === 'networkRequest' && Array.isArray(res.errors)) {
-		const props = {
-			request: path(['request', 'props'], data) || [],
-			response: path(['response', 'props'], data) || [],
-		};
-
-		// go through errors array
-		for (let i = 0; i < res.errors.length; i++) {
-			const error = res.errors[i];
-			let expected, actual, propName;
-
-			if (error.type === 'noUriFound') {
-				expected = path(['comparator', 'val'], data);
-				actual = t['actualExpected.requestNotMade']();
-				propName = 'url';
-				values.push(valTemplate(`${propName}: ${expected}`, false));
-				values.push(valTemplate(`${propName}: ${actual}`, true));
-				continue;
-			}
-
-			const nameComparator = error.type === 'header' ? error.name : NETWORK_PROP['@' + error.type];
-			const propsField = props[error.message].find(field => field.name === nameComparator);
-
-			if (propsField) {
-				expected = error.type === 'method' ? NETWORK_METHOD[propsField.val] : propsField.val;
-				actual = 'actualValue' in error
-					? error.actualValue
-					: error.reason === 'notFound' ? t['actualExpected.notFound']() : t['actualExpected.unknown']();
-				propName = `${error.message} ${error.type}` + (error.name ? ` "${error.name}"` : '');
-				values.push(valTemplate(`${propName}: ${expected}`, false));
-				values.push(valTemplate(`${propName}: ${actual}`, true));
-			}
-		}
+		values.push(...getNetworkActualExpected(res, data));
 	}
 
 	return values.length ? `${EOL}\t${t['actualExpected.failingChecks']()}${arrToString(values, true)}` : '';
