+++ lib/utils/testHelpers/testInputError.js
@@ -1,21 +1,17 @@
 /* istanbul ignore file */
 
+const {curry} = require('ramda');
 const assert = require('assert');
 const assertThrowsAsync = require('../assertThrowsAsync');
 const SuitestError = require('../SuitestError');
 
-module.exports.testInputErrorAsync = (fn, args = [], expect = {}, msg) => {
-	return assertThrowsAsync(fn.bind(fn, ...args), {
+const commonTestInputError = curry((asserter, fn, args = [], expect = {}, msg) => {
+	return asserter(fn.bind(fn, ...args), {
 		type: SuitestError.type,
 		code: SuitestError.INVALID_INPUT,
 		...expect,
 	}, msg || `Invalid error if ${args.length ? args.map(JSON.stringify).join(', ') : 'no args'}`);
-};
+});
 
-module.exports.testInputErrorSync = (fn, args = [], expect = {}, msg) => {
-	return assert.throws(fn.bind(fn, ...args), {
-		type: SuitestError.type,
-		code: SuitestError.INVALID_INPUT,
-		...expect,
-	}, msg || `Invalid error if ${args.length ? args.map(JSON.stringify).join(', ') : 'no args'}`);
-};
+module.exports.testInputErrorAsync = commonTestInputError(assertThrowsAsync);
+module.exports.testInputErrorSync = commonTestInputError(assert.throws);
