+++ lib/validataion/validators.js
@@ -30,49 +30,58 @@ function throwError(text) {
 
 const getElementPropName = (mapOfNames, name) => typeof mapOfNames[name] === 'string' ? mapOfNames[name] : name;
 
-function prettifyJsonSchemaErrors(validate) {
-	let errors = [];
+function prettifyElementPropsErrors(validate) {
 	const getDataName = path(['data', 'name']);
 
+	return validate.errors
+		.filter(getDataName)
+		.map(i => {
+			const dataName = getDataName(i);
+
+			if (dataName in ELEMENT_PROP_M) {
+				const elementPropTye = typeof dataName === 'string' ?
+					ELEMENT_PROP_TYPES[ELEMENT_PROP_M[dataName]] :
+					ELEMENT_PROP_TYPES[dataName];
+				const elementProp = getElementPropName(ELEMENT_PROP_M, dataName);
+
+				return `Element property '${elementProp}' should be of type ${elementPropTye}`;
+			}
+
+			return `Element property ${dataName.toString()} is unknown Symbol.`;
+		});
+}
+
+function prettifyElementSelectorsErrors(validate) {
+	return validate.errors.map(error => {
+		if (error.dataPath === '.index') {
+			return incorrectElementIndex();
+		}
+		// check if element selector contains additional properties
+		if (error.keyword === 'additionalProperties') {
+			const properties = Object.keys(error.parentSchema.properties);
+			const additionalProperties = Object.keys(error.data)
+				.filter(prop => !properties.includes(prop))
+				.join(', ');
+
+			// will be something like:
+			// Element selector should NOT have additional properties: otherProp, otherProp2
+			return `${error.message}: ${additionalProperties || error.params.additionalProperty}`;
+		}
+
+		return error.message;
+	});
+}
+
+function prettifyJsonSchemaErrors(validate) {
+	let errors = [];
+
 	if (validate.schema.schemaId === validationKeys.ELEMENT_PROPS) {
-		errors = validate.errors
-			.filter(getDataName)
-			.map(i => {
-				const dataName = getDataName(i);
-
-				if (dataName in ELEMENT_PROP_M) {
-					const elementPropTye = typeof dataName === 'string' ?
-						ELEMENT_PROP_TYPES[ELEMENT_PROP_M[dataName]] :
-						ELEMENT_PROP_TYPES[dataName];
-					const elementProp = getElementPropName(ELEMENT_PROP_M, dataName);
-
-					return `Element property '${elementProp}' should be of type ${elementPropTye}`;
-				}
-
-				return `Element property ${dataName.toString()} is unknown Symbol.`;
-			});
+		errors = prettifyElementPropsErrors(validate);
 	} else if (validate.schema.schemaId === validationKeys.CONFIGURE) {
 		// example: Invalid input provided for configuration object. 'disallowCrashReports' should be boolean.
 		errors = validate.errors.map(i => `'${i.dataPath.slice(1)}' ${i.message}.`);
 	} else if (validate.schema.schemaId === validationKeys.ELEMENT_SELECTOR) {
-		errors = validate.errors.map(error => {
-			if (error.dataPath === '.index') {
-				return incorrectElementIndex();
-			}
-			// check if element selector contains additional properties
-			if (error.keyword === 'additionalProperties') {
-				const properties = Object.keys(error.parentSchema.properties);
-				const additionalProperties = Object.keys(error.data)
-					.filter(prop => !properties.includes(prop))
-					.join(', ');
-
-				// will be something like:
-				// Element selector should NOT have additional properties: otherProp, otherProp2
-				return `${error.message}: ${additionalProperties || error.params.additionalProperty}`;
-			}
-
-			return error.message;
-		});
+		errors = prettifyElementSelectorsErrors(validate);
 	} else {
 		errors = validate.errors.map(i => i.message);
 	}
