+++ testDefinition/examplesTS/executeCommandChain.ts
@@ -1,0 +1,14 @@
+import * as suitest from '../../index';
+
+async function executeCommandTest() {
+	const {executeCommand} = suitest;
+
+	await executeCommand('adasd');
+	await executeCommand(() => '');
+	await executeCommand((a: string) => a.toLowerCase());
+
+	executeCommand(() => {}).abandon();
+	executeCommand(() => {}).then();
+	executeCommand(() => {}).clone();
+	executeCommand(() => {}).toString();
+}
